# Remove some branches from number parsing