This simplifies integer parsing (and float parsing, a little bit), in particular removing a bunch of visible branches. (I say visible because I believe the compiler was getting rid of said branches given that performance and branch misses don't really change.) It does not change performance at all on Skylake / GCC 9.

The main changes:

1. **parse_digit:** Moved digit parsing to a common parse_digit() function, eliminating is_integer and doing the comparison against the parsed int. The compiler was already doing this for us, but this both makes the callers easier to read, and makes the intended flow clearer instead of having redundant operations in the code to be optimized.
2. **Remove `case '-'`:** Removed the `found_minus` parameter and the special cases for '-' in the structural_parser. This makes no performance difference on skylake, indicating it's an optimization with no benefit. Either (a) the compiler was already combining a lot of the inline code, or (b) the , and I don't know that we want it to be.
3. **Move error checks:** Instead of many error check branches, we consolidate error checks for each branch (int, dec, exp) to a single if() after the parse. The de-branching was likely already being done by the compiler, given the lack of performance improvement, but this gives the user less code to read and the compiler fewer decisions to make. This also simplifies the main body of the parse and makes it easier to follow, IMO.
4. **Remove redundant check:** Removed the weird double `if (is_integer)` we talked about in #990. I suspect the many if statements were the reason earlier attempts didn't take--it probably pushed some of the compiler's debranching optimizations over the edge. My experiments tell me the compiler is really not very good at debranching things by itself.
5. **Move exp overflow check:** Exponent overflow was being checked after each digit. Not a huge deal, but there is no reason to do this if you can check after (as with int parsing).