@jkeiser 2020-07-01
This came largely as a result of me experimenting to try to get rid of that redundant if.