@jkeiser 2020-07-02
@lemire I was able to use parse_digit in float and exponent as well, figured out the issues there. Patch is now number parsing all up.