@jkeiser 2020-07-02
skylake GCC 9:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| update-center                    |   8330 |  43.3 |  42.9 |    0% | 142.1 | 142.1 |    0% |   2564 |   2372 |    7% |      0 |      0 |    0% |  25147 |  25179 |    0% |
| mesh                             |  11306 |  45.7 |  45.4 |    0% | 174.2 | 174.2 |    0% |    732 |    603 |   17% |      3 |      0 |    0% |  43116 |  43031 |    0% |
| numbers                          |   2345 |  38.4 |  38.3 |    0% | 144.3 | 144.3 |    0% |    104 |     94 |    9% |      0 |      0 |    0% |   6932 |   6581 |    5% |
| twitter                          |   9867 |  42.9 |  42.8 |    0% | 148.8 | 148.8 |    0% |   1728 |   1680 |    2% |      0 |      0 |    0% |  27142 |  27158 |    0% |
| mesh.pretty                      |  24646 |  35.7 |  35.7 |    0% | 137.7 | 137.7 |    0% |     45 |     27 |   40% |    571 |    120 |   78% |  69507 |  69316 |    0% |
| marine_ik                        |  46616 |  53.3 |  53.2 |    0% | 180.9 | 180.9 |    0% |   8787 |   8715 |    0% |  66984 |  75403 |  -12% | 186268 | 186711 |    0% |
| instruments                      |   3442 |  37.7 |  37.6 |    0% | 143.1 | 143.1 |    0% |     93 |     91 |    2% |      0 |      0 |    0% |  10640 |   9899 |    6% |
| random                           |   7976 |  57.1 |  57.2 |    0% | 209.3 | 209.3 |    0% |   1018 |   1021 |    0% |      0 |      0 |    0% |  27671 |  27736 |    0% |
| apache_builds                    |   1988 |  36.6 |  36.6 |    0% | 139.3 | 139.3 |    0% |     16 |     21 |  -31% |      0 |      0 |    0% |   4668 |   5080 |   -8% |
| github_events                    |   1017 |  34.0 |  34.1 |    0% | 129.8 | 129.8 |    0% |      8 |      5 |   37% |      0 |      0 |    0% |    893 |   1670 |  -87% |
| citm_catalog                     |  26987 |  34.0 |  34.0 |    0% | 130.7 | 130.7 |    0% |    161 |    177 |   -9% |    193 |    475 | -146% |  71773 |  71974 |    0% |
| gsoc-2018                        |  51997 |  35.8 |  36.0 |    0% | 101.1 | 101.1 |    0% |  15042 |  15019 |    0% |  33419 |  25467 |   23% | 117537 | 116614 |    0% |
| twitterescaped                   |   8787 |  40.8 |  41.1 |    0% | 138.1 | 138.1 |    0% |   1808 |   1944 |   -7% |      0 |      0 |    0% |  25126 |  25010 |    0% |
| canada                           |  35172 |  44.5 |  45.0 |   -1% | 157.6 | 157.6 |    0% |   5326 |   5336 |    0% |   8054 |  11433 |  -41% | 115097 | 115614 |    0% |