@jkeiser 2020-07-02
NOTE: I added parse_digit to decimal and exponent parsing as well. I had to add NO_SANITIZE_UNDEFINED to parse_digit to make it work, because the fuzzer passes `-0e000000000000000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000102339566` to it and it temporarily overflows (we check it after the parse instead of during the parse now). I'm a little surprised the fuzzer doesn't pass similar stuff to integer/decimal parsing, but I don't know yet how we guide the fuzzer so coudn't check.