@jkeiser 2020-07-10
I think [Use parse_digit for decimal and exp parsing as well](https://github.com/simdjson/simdjson/pull/1018/commits/c1c274bee980d4ff6e632c52e112ac26db6135cf) may have introduced a minor westmere-only performance regression. Tweaking.