@aqrit 2020-07-10
> 2. Remove case '-': 

The switch/case in `parse_value()` and `parse_structurals()` are implemented as binary decision trees in gcc.
I think this may obscure any performance differences.

Would a jump table implementation be faster?... Note: `c >> 3` is a perfect hash!

```
// part of the binary decision tree
	movzx   edx, byte ptr [r13+rax+0]
	cmp     dl, 66h ; 'f'
	jz      loc_B7E8
	ja      short loc_AFF0
	cmp     dl, 2Dh ; '-'
	jz      loc_AEC4
	ja      loc_AEB8
	cmp     dl, 22h ; '"'
	jnz     short loc_AF9E
	mov     dword ptr [rbx+58h], 5
	mov     eax, 5
	jmp     loc_AB32
loc_AFF0:
	cmp     dl, 6Eh ; 'n'
	jz      loc_B7D0
	cmp     dl, 74h ; 't'
	jnz     short loc_AF9E
```