@jkeiser 2020-07-11
Holy crap, I was assuming it was a jump table, especially since we have locations for each thing to jump to!