@jkeiser 2020-07-11
After factoring though, it stands to reason that there wouldn't be much performance difference from having a separate - branch. The places where it matters are few and far between, and largely capable of being branchless. There is a LOT of code in parse_number, as well, which, if not de-duplicated, would increase instruction cache / fetch load as well as BPU load. And the fact that removing it didn't change instruction count indicates it's probably de-duped.

Also, I've attempted to add template<bool found_minus> to parse_number before, and got a throughput decrease. I don't know the source though, it's possible I did something boneheaded, even if it wasn't a very big change.