@lemire 2020-07-14
To people reading this: there are perf. regression warnings. I have examined the performance and so has @jkeiser. It is not clear whether this PR actually brings all around performance gains, but I expect that it might (despite the warnings).

Even so, this PR needs to be merged on software engineering grounds. It makes the code nicer.